gaerangmari/pyproject.toml
[tool.poetry]
name = "gaerangmari"
version = "0.1.0"
description = "강아지 산책 커뮤니티 프로젝트"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.11"
isort = "^5.12"
flake8-pyproject = "^1.2.3"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
django-debug-toolbar = "^4.4.6"

[tool.bandit]
targets = ["gaerangmari"]
exclude_dirs = ["tests", "venv"]
skips = ["B311"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88